app: extract App.tsx pre-component layer into lib/ and components/ modules
Mechanical decomposition of the 5,278-line App.tsx: the 1,135 lines ahead
of the App() component move verbatim into

- src/lib/config.ts        network/profile constants, storage keys, role meta
- src/lib/appTypes.ts      log/activity/alert type definitions
- src/lib/format.ts        pure formatting/validation/parsing helpers
- src/lib/logs.tsx         log classification, highlighting, parsing
- src/lib/connections.ts   connection normalization, ports, default profiles
- src/components/          AppIcon, EmptyStateIcon, ScrollToTop

App() itself is unchanged and now imports these modules; declaration order
and logic are preserved. check-ui-guards now scans the combined renderer
sources (App.tsx + src/lib + src/components) so its negative assertions
keep their strength across the decomposition.

Gates: tsc typecheck, vitest (37), check:ui-guards, electron-vite build,
and the packaged-app autostart guard all pass.
