+++ hegemon-app/src/App.tsx
@@ -19,1140 +19,122 @@ import blockMinedAudio from './assets/sounds/block-mined.wav';
 import blockReceivedAudio from './assets/sounds/block-received.wav';
 import { computeNodeDisplayState, legacyContactWarning } from './appGuards';
 
-const defaultStorePath = '~/.hegemon-wallet-010';
-const canonicalTestnetP2pPort = 30333;
-const shieldedAddressPrefix = 'shca1';
-const shieldedAddressLength = 2634;
-const shieldedAddressDataCharset = /^[023456789acdefghjklmnpqrstuvwxyz]+$/;
-const shieldedAddressSeparatorPattern = /[\s\u200B\u200C\u200D\uFEFF]+/g;
-const approvedSeedEntries = ['hegemon.pauli.group:30333', 'devnet.hegemonprotocol.com:30333'] as const;
-const approvedSeeds = approvedSeedEntries.join(',');
-const hegemonNetworkName = 'Hegemon';
-const hegemonNetworkVersionLabel = 'Hegemon 0.10';
-const defaultDevConnectionLabel = hegemonNetworkName;
-const legacyDefaultConnectionLabels = new Set(['Local node', 'Native 0.10 devnet', hegemonNetworkName]);
-const legacyHegemonConnectionLabels = new Set([
-  'hegemon-dev',
-  'hegemon-dev P2P 0.10',
-  'Hegemon Native Dev',
-  'Hegemon Native Devnet',
-  'Native 0.10 devnet'
-]);
-const defaultDevBasePath = '~/.hegemon-node-native-010-dev';
-const legacyDesktopRpcPort = 9944;
-const legacySeedAliases: Record<string, string> = {
-  'hegemon.pauli.group:31333': approvedSeeds,
-  'hegemon.pauli.group:30333': approvedSeeds,
-  '158.69.222.121:31333': approvedSeeds,
-  '158.69.222.121:30333': approvedSeeds,
-  'devnet.hegemonprotocol.com:30333': approvedSeeds,
-  '51.222.86.107:30333': approvedSeeds
-};
-const connectionsKey = 'hegemon.nodeConnections';
-const activeConnectionKey = 'hegemon.activeConnection';
-const walletConnectionKey = 'hegemon.walletConnection';
-const walletStorePathKey = 'hegemon.walletStorePath';
-const walletAutoLockEnabledKey = 'hegemon.walletAutoLockEnabled';
-const walletAutoLockMinutesKey = 'hegemon.walletAutoLockMinutes';
-const blockAlertEnabledKey = 'hegemon.blockAlertEnabled';
-const minWalletPassphraseLength = 12;
-const defaultRpcPort = 9955;
-const defaultP2pPort = canonicalTestnetP2pPort;
-const maxDesktopMineThreads = 4;
-const walletHistoryPageSize = 24;
-const defaultMineThreads = (() => {
-  const hardwareConcurrency =
-    typeof navigator !== 'undefined' && Number.isFinite(navigator.hardwareConcurrency)
-      ? Number(navigator.hardwareConcurrency)
-      : 1;
-  const target = Math.floor(hardwareConcurrency / 4);
-  return Math.max(1, Math.min(maxDesktopMineThreads, target || 1));
-})();
 
-const participationRoleLabels: Record<NodeParticipationRole, string> = {
-  full_node: 'Relay node',
-  authoring_pool: 'Mining node'
-};
-
-const connectionModeLabels: Record<NodeConnectionMode, string> = {
-  local: 'Managed local',
-  remote: 'Local RPC endpoint'
-};
-
-const participationRoleMeta: Record<
-  NodeParticipationRole,
-  {
-    statusLabel?: string;
-    statusTone: 'ok' | 'warn';
-    summary: string;
-    guidance: string;
-  }
-> = {
-  full_node: {
-    statusTone: 'ok',
-    summary: 'Verifies the network, serves wallet traffic, and relays chain state without local block production.',
-    guidance:
-      'Use this for wallets, verification, and relaying. Switch to Mining node only if this machine should build and mine blocks.'
-  },
-  authoring_pool: {
-    statusLabel: 'Operator only',
-    statusTone: 'warn',
-    summary:
-      'Runs the public block author. This node accepts proof-ready transactions, builds the commitment proof, mines locally, and broadcasts final blocks.',
-    guidance:
-      'Use this only on the machine that should mine blocks. Everyone else should stay on Relay node.'
-  }
-};
-
-const inferParticipationRole = (connection: NodeConnection): NodeParticipationRole => {
-  if (connection.participationRole === 'full_node' || connection.participationRole === 'authoring_pool') {
-    return connection.participationRole;
-  }
-  return connection.miningIntent || connection.minerAddress ? 'authoring_pool' : 'full_node';
-};
-
-const normalizeTxId = (value: string | null | undefined) => {
-  if (!value) {
-    return null;
-  }
-  const trimmed = value.trim();
-  if (!trimmed) {
-    return null;
-  }
-  return trimmed.replace(/^0x/i, '').toLowerCase();
-};
-
-const canonicalizeSeedEntry = (seed: string) => {
-  const normalized = seed.trim().toLowerCase();
-  if (!normalized) {
-    return '';
-  }
-  return legacySeedAliases[normalized] ?? normalized;
-};
-
-const normalizeSeedsValue = (value: string | null | undefined) => {
-  const normalized: string[] = [];
-  const seen = new Set<string>();
-  for (const rawSeed of (value ?? '').split(',')) {
-    for (const seed of canonicalizeSeedEntry(rawSeed).split(',')) {
-      if (!seed || seen.has(seed)) {
-        continue;
-      }
-      seen.add(seed);
-      normalized.push(seed);
-    }
-  }
-  return normalized.join(',');
-};
-
-const normalizeNetworkDisplayName = (value: string | null | undefined) => {
-  const trimmed = value?.trim();
-  if (!trimmed) {
-    return hegemonNetworkName;
-  }
-  const lower = trimmed.toLowerCase();
-  if (
-    lower === 'hegemon network' ||
-    lower === 'hegemon native' ||
-    lower === 'hegemon native dev' ||
-    lower === 'native 0.10 devnet' ||
-    lower === 'hegemon-dev' ||
-    lower.startsWith('hegemon native ')
-  ) {
-    return hegemonNetworkName;
-  }
-  return trimmed;
-};
-
-const makeId = () => {
-  if (typeof crypto !== 'undefined' && 'randomUUID' in crypto) {
-    return crypto.randomUUID();
-  }
-  return `conn-${Math.random().toString(36).slice(2, 10)}`;
-};
-
-const clampAutoLockMinutes = (value: number) => Math.min(Math.max(value, 1), 120);
-
-type LogLevel = 'info' | 'warn' | 'error' | 'debug';
-
-type LogCategory = 'mining' | 'sync' | 'network' | 'consensus' | 'storage' | 'rpc' | 'other';
-
-type LogEntry = {
-  id: string;
-  timestamp: string | null;
-  level: LogLevel;
-  category: LogCategory;
-  message: string;
-  raw: string;
-  highlight?: string;
-};
-
-type ActivityStatus = 'processing' | 'pending' | 'confirmed' | 'failed';
-
-type SendAttempt = {
-  id: string;
-  storePath: string;
-  createdAt: string;
-  recipient: string;
-  amount: number;
-  fee: number;
-  memo?: string;
-  status: ActivityStatus;
-  txId?: string;
-  error?: string;
-  notesNeeded?: number;
-  walletNoteCount?: number;
-  maxInputs?: number;
-  consolidationExpected?: number;
-  consolidationExpectedBlocks?: number;
-};
-
-type ActivityStep = {
-  id: string;
-  label: string;
-  status: ActivityStatus;
-  txId?: string;
-  confirmations?: number;
-};
-
-type ActivityEntry = {
-  id: string;
-  source: 'attempt' | 'wallet';
-  createdAt: string;
-  recipient: string;
-  amount: number;
-  fee: number;
-  memo?: string;
-  status: ActivityStatus;
-  txId?: string;
-  confirmations?: number;
-  error?: string;
-  notesNeeded?: number;
-  walletNoteCount?: number;
-  maxInputs?: number;
-  consolidationExpected?: number;
-  consolidationExpectedBlocks?: number;
-  consolidationSubmitted?: number;
-  consolidationConfirmed?: number;
-  steps?: ActivityStep[];
-};
-
-type DisclosureGroup = {
-  txId: string;
-  createdAt: string;
-  outputs: WalletDisclosureRecord[];
-};
-
-type NodeTransitionAction = 'starting' | 'stopping';
-
-const isWalletSessionClosedError = (message: string) => {
-  const normalized = message.toLowerCase();
-  return (
-    normalized.includes('walletd process not running') ||
-    normalized.includes('walletd stopped') ||
-    normalized.includes('walletd stdin not available') ||
-    normalized.includes('wallet is locked') ||
-    normalized.includes('wallet unlock token') ||
-    normalized.includes('expired')
-  );
-};
-
-type NodeTransition = {
-  action: NodeTransitionAction;
-  connectionId: string;
-  startedAt: number;
-};
-
-type UiTone = 'ok' | 'warn' | 'error' | 'neutral';
-type BlockAlertTone = 'self' | 'other';
-type BlockAlertStep = { frequency: number; duration: number; gap?: number };
-type EmptyStateIconName = 'terminal' | 'transactions' | 'contacts' | 'disclosure';
-type AppIconName =
-  | 'overview'
-  | 'node'
-  | 'wallet'
-  | 'send'
-  | 'disclosure'
-  | 'console'
-  | 'height'
-  | 'target'
-  | 'sync'
-  | 'key'
-  | 'peers'
-  | 'mining'
-  | 'endpoint';
-
-const logCategoryOrder: LogCategory[] = ['mining', 'sync', 'network', 'consensus', 'storage', 'rpc', 'other'];
-
-const logCategoryLabels: Record<LogCategory, string> = {
-  mining: 'Mining',
-  sync: 'Sync',
-  network: 'Network',
-  consensus: 'Consensus',
-  storage: 'Storage',
-  rpc: 'RPC',
-  other: 'Other'
-};
-
-const toBaseUnits = (value: string) => {
-  const parsed = Number.parseFloat(value);
-  if (Number.isNaN(parsed) || !Number.isFinite(parsed)) {
-    return null;
-  }
-  return Math.round(parsed * 100_000_000);
-};
-
-const formatNumber = (value: number | null | undefined) => {
-  if (value === null || value === undefined) {
-    return 'N/A';
-  }
-  return value.toLocaleString();
-};
-
-const formatHgm = (value: number) => `${(value / 100_000_000).toFixed(8)} HGM`;
-
-const formatBlockCount = (value: number) => `${formatNumber(value)} ${value === 1 ? 'block' : 'blocks'}`;
-
-const AppIcon = ({ name }: { name: AppIconName }) => {
-  const sharedProps = {
-    viewBox: '0 0 24 24',
-    fill: 'none',
-    'aria-hidden': true
-  };
-
-  const paths: Record<AppIconName, JSX.Element> = {
-    overview: (
-      <>
-        <path d="M3.5 11.5 12 4l8.5 7.5" />
-        <path d="M5.5 10.5v8h13v-8" />
-        <path d="M9.5 18.5v-5h5v5" />
-      </>
-    ),
-    node: (
-      <>
-        <path d="M12 4.5v5" />
-        <path d="M6.5 14.5h11" />
-        <path d="M6.5 14.5v5" />
-        <path d="M17.5 14.5v5" />
-        <path d="M9.5 9.5h5v5h-5z" />
-        <path d="M4.5 19.5h4" />
-        <path d="M15.5 19.5h4" />
-      </>
-    ),
-    wallet: (
-      <>
-        <path d="M4 7.5h14.5a2 2 0 0 1 2 2v8a2 2 0 0 1-2 2H5.5A2.5 2.5 0 0 1 3 17V8.5a1 1 0 0 1 1-1Z" />
-        <path d="M4.5 7.5 16 4.5" />
-        <path d="M16.5 13.5h4" />
-      </>
-    ),
-    send: (
-      <>
-        <path d="M4 5.5 20 12 4 18.5l3-6.5-3-6.5Z" />
-        <path d="M7 12h7" />
-      </>
-    ),
-    disclosure: (
-      <>
-        <path d="M6.5 3.5h8l4 4v13h-12z" />
-        <path d="M14.5 3.5v4h4" />
-        <path d="M8.5 14c1.2-1.7 2.4-2.5 3.5-2.5s2.3.8 3.5 2.5c-1.2 1.7-2.4 2.5-3.5 2.5S9.7 15.7 8.5 14Z" />
-        <path d="M11.25 14a.75.75 0 1 0 1.5 0 .75.75 0 0 0-1.5 0Z" />
-      </>
-    ),
-    console: (
-      <>
-        <path d="M4.5 6.5h15v11h-15z" />
-        <path d="m7 10 2.5 2L7 14" />
-        <path d="M12 14h4" />
-      </>
-    ),
-    height: (
-      <>
-        <path d="M4 16.5 8.5 12l3 3 8-8" />
-        <path d="M15.5 7.5h4v4" />
-      </>
-    ),
-    target: (
-      <>
-        <path d="M12 4v3" />
-        <path d="M12 17v3" />
-        <path d="M4 12h3" />
-        <path d="M17 12h3" />
-        <path d="M7.5 12a4.5 4.5 0 1 0 9 0 4.5 4.5 0 0 0-9 0Z" />
-      </>
-    ),
-    sync: (
-      <>
-        <path d="M18.5 8.5A7 7 0 0 0 6 7l-1.5 2.5" />
-        <path d="M4.5 7.5v2h2" />
-        <path d="M5.5 15.5A7 7 0 0 0 18 17l1.5-2.5" />
-        <path d="M19.5 16.5v-2h-2" />
-      </>
-    ),
-    key: (
-      <>
-        <path d="M4.5 14.5a4 4 0 1 0 3-3.9" />
-        <path d="M11.5 12.5 20 4" />
-        <path d="M16.5 7.5 18 9" />
-        <path d="M14.5 9.5 16 11" />
-      </>
-    ),
-    peers: (
-      <>
-        <path d="M8 10a2.5 2.5 0 1 0 0-5 2.5 2.5 0 0 0 0 5Z" />
-        <path d="M16 10a2.5 2.5 0 1 0 0-5 2.5 2.5 0 0 0 0 5Z" />
-        <path d="M4.5 19c.7-3.4 2-5 3.5-5s2.8 1.6 3.5 5" />
-        <path d="M12.5 19c.7-3.4 2-5 3.5-5s2.8 1.6 3.5 5" />
-      </>
-    ),
-    mining: (
-      <>
-        <path d="m5 19 9.5-9.5" />
-        <path d="M12 7.5 16.5 3 21 7.5l-4.5 4.5" />
-        <path d="M8.5 15.5 10 17" />
-      </>
-    ),
-    endpoint: (
-      <>
-        <path d="M6.5 7.5h11v9h-11z" />
-        <path d="M9 20h6" />
-        <path d="M12 16.5V20" />
-        <path d="M4 10h2.5" />
-        <path d="M17.5 10H20" />
-      </>
-    )
-  };
-
-  return (
-    <svg className={`app-icon app-icon-${name}`} {...sharedProps}>
-      {paths[name]}
-    </svg>
-  );
-};
-
-const EmptyStateIcon = ({ name }: { name: EmptyStateIconName }) => {
-  const sharedProps = {
-    className: `empty-state-glyph ${name}`,
-    viewBox: '0 0 48 48',
-    fill: 'none',
-    'aria-hidden': true
-  };
-
-  if (name === 'terminal') {
-    return (
-      <svg {...sharedProps}>
-        <path d="M11 14.5h26a3 3 0 0 1 3 3v17a3 3 0 0 1-3 3H11a3 3 0 0 1-3-3v-17a3 3 0 0 1 3-3Z" />
-        <path d="M15 22.5 20 27l-5 4.5" />
-        <path d="M24 31.5h9" />
-      </svg>
-    );
-  }
-
-  if (name === 'contacts') {
-    return (
-      <svg {...sharedProps}>
-        <path d="M13 10.5h18a4 4 0 0 1 4 4v19a4 4 0 0 1-4 4H13a3 3 0 0 1-3-3v-21a3 3 0 0 1 3-3Z" />
-        <path d="M35 17h4M35 24h4M35 31h4" />
-        <path d="M18 22.5a4 4 0 1 0 8 0 4 4 0 0 0-8 0Z" />
-        <path d="M15.5 32.5c1.7-3.1 4.1-4.7 6.5-4.7s4.8 1.6 6.5 4.7" />
-      </svg>
-    );
-  }
-
-  if (name === 'disclosure') {
-    return (
-      <svg {...sharedProps}>
-        <path d="M14 8.5h14l8 8v20a3 3 0 0 1-3 3H14a3 3 0 0 1-3-3v-25a3 3 0 0 1 3-3Z" />
-        <path d="M28 8.5v8h8" />
-        <path d="M16.5 27c2.3-3.1 4.8-4.7 7.5-4.7s5.2 1.6 7.5 4.7c-2.3 3.1-4.8 4.7-7.5 4.7s-5.2-1.6-7.5-4.7Z" />
-        <path d="M22 27a2 2 0 1 0 4 0 2 2 0 0 0-4 0Z" />
-      </svg>
-    );
-  }
-
-  return (
-    <svg {...sharedProps}>
-      <path d="M12 18.5h20" />
-      <path d="m30 13.5 5 5-5 5" />
-      <path d="M36 29.5H16" />
-      <path d="m18 24.5-5 5 5 5" />
-      <path d="M13 11.5h9" opacity="0.55" />
-      <path d="M26 36.5h9" opacity="0.55" />
-    </svg>
-  );
-};
-
-const normalizeShieldedAddressInput = (value: string) =>
-  value.replace(shieldedAddressSeparatorPattern, '').trim().toLowerCase();
-
-const validateShieldedAddressInput = (value: string, label = 'Address') => {
-  const normalized = normalizeShieldedAddressInput(value);
-  if (!normalized) {
-    return `${label} is required.`;
-  }
-  if (!normalized.startsWith(shieldedAddressPrefix)) {
-    return `${label} must start with ${shieldedAddressPrefix}.`;
-  }
-  if (normalized.length !== shieldedAddressLength) {
-    return `${label} looks truncated or corrupt. Expected ${shieldedAddressLength} chars, got ${normalized.length}.`;
-  }
-  const payload = normalized.slice(shieldedAddressPrefix.length);
-  if (!shieldedAddressDataCharset.test(payload)) {
-    return `${label} contains unsupported bech32 characters.`;
-  }
-  return null;
-};
-
-const formatAddress = (address: string) => {
-  const normalized = normalizeShieldedAddressInput(address);
-  if (normalized.length <= 28) {
-    return normalized || address;
-  }
-  const middleStart = Math.max(0, Math.floor(normalized.length / 2) - 4);
-  return `${normalized.slice(0, 10)}...${normalized.slice(middleStart, middleStart + 8)}...${normalized.slice(-8)}`;
-};
-
-const formatCompactPath = (value: string | null | undefined) => {
-  const trimmed = value?.trim();
-  if (!trimmed) {
-    return 'N/A';
-  }
-  if (trimmed.length <= 48) {
-    return trimmed;
-  }
-  const normalized = trimmed.replace(/\\/g, '/');
-  const parts = normalized.split('/').filter(Boolean);
-  if (parts.length <= 2) {
-    return `${trimmed.slice(0, 20)}...${trimmed.slice(-20)}`;
-  }
-  const prefix = normalized.startsWith('/') ? '/' : '';
-  return `${prefix}${parts[0]}/.../${parts.slice(-2).join('/')}`;
-};
-
-const formatEndpoint = (value: string | null | undefined) => {
-  const trimmed = value?.trim();
-  if (!trimmed) {
-    return 'N/A';
-  }
-  if (trimmed.length <= 40) {
-    return trimmed;
-  }
-  try {
-    const parsed = new URL(trimmed);
-    return `${parsed.protocol}//${parsed.host}${parsed.pathname === '/' ? '' : parsed.pathname}`;
-  } catch {
-    return `${trimmed.slice(0, 18)}...${trimmed.slice(-18)}`;
-  }
-};
-
-const formatSeedList = (value: string[] | string | null | undefined) => {
-  const seeds = Array.isArray(value) ? value : (value ?? '').split(',');
-  const normalized = seeds.map((seed) => seed.trim()).filter(Boolean);
-  return normalized.length ? normalized.join(', ') : 'N/A';
-};
-
-const humanizeWalletAddressError = (error: unknown) => {
-  const message = error instanceof Error ? error.message : String(error);
-  if (message.toLowerCase().includes('invalid address encoding')) {
-    const lengthMatch = message.match(/invalid address length: expected (\d+) bytes, got (\d+)/i);
-    if (lengthMatch) {
-      const [, expectedBytes, actualBytes] = lengthMatch;
-      return `Recipient address is malformed or truncated. Expected ${expectedBytes} address bytes, got ${actualBytes}. Ask your contact to re-copy their full shielded address.`;
-    }
-    return 'Recipient address is malformed or truncated. Ask your contact to re-copy their full shielded address.';
-  }
-  return message;
-};
-
-const formatBytes = (value: number | null | undefined) => {
-  if (value === null || value === undefined || Number.isNaN(value)) {
-    return 'N/A';
-  }
-  if (value < 1024) {
-    return `${value} B`;
-  }
-  const units = ['KB', 'MB', 'GB', 'TB'];
-  let remaining = value;
-  let unitIndex = -1;
-  while (remaining >= 1024 && unitIndex < units.length - 1) {
-    remaining /= 1024;
-    unitIndex += 1;
-  }
-  return `${remaining.toFixed(2)} ${units[unitIndex]}`;
-};
-
-const formatHashRate = (value: number | null | undefined) => {
-  if (value === null || value === undefined || Number.isNaN(value)) {
-    return 'N/A';
-  }
-  const units = ['H/s', 'KH/s', 'MH/s', 'GH/s', 'TH/s'];
-  let remaining = value;
-  let unitIndex = 0;
-  while (remaining >= 1000 && unitIndex < units.length - 1) {
-    remaining /= 1000;
-    unitIndex += 1;
-  }
-  const decimals = remaining >= 100 ? 0 : remaining >= 10 ? 1 : 2;
-  return `${remaining.toFixed(decimals)} ${units[unitIndex]}`;
-};
-
-const formatDuration = (seconds: number | null | undefined) => {
-  if (seconds === null || seconds === undefined || Number.isNaN(seconds)) {
-    return 'N/A';
-  }
-  const safeSeconds = Math.max(0, Math.floor(seconds));
-  const days = Math.floor(safeSeconds / 86400);
-  const hours = Math.floor((safeSeconds % 86400) / 3600);
-  const minutes = Math.floor((safeSeconds % 3600) / 60);
-  const secs = safeSeconds % 60;
-  const parts: string[] = [];
-  if (days) {
-    parts.push(`${days}d`);
-  }
-  if (hours || days) {
-    parts.push(`${hours}h`);
-  }
-  if (minutes || hours || days) {
-    parts.push(`${minutes}m`);
-  }
-  parts.push(`${secs}s`);
-  return parts.join(' ');
-};
-
-const formatHash = (value: string | null | undefined) => {
-  if (!value) {
-    return 'N/A';
-  }
-  if (value.length <= 20) {
-    return value;
-  }
-  return `${value.slice(0, 10)}...${value.slice(-8)}`;
-};
-
-const buildBlockAlertPattern = (tone: BlockAlertTone): BlockAlertStep[] => {
-  if (tone === 'self') {
-    return [
-      { frequency: 1480, duration: 0.08, gap: 0.05 },
-      { frequency: 1760, duration: 0.08, gap: 0.05 },
-      { frequency: 2090, duration: 0.12 }
-    ];
-  }
-  return [
-    { frequency: 330, duration: 0.22, gap: 0.08 },
-    { frequency: 220, duration: 0.26 }
-  ];
-};
-
-const ScrollToTop = () => {
-  const location = useLocation();
-
-  useEffect(() => {
-    window.scrollTo({ top: 0, left: 0, behavior: 'auto' });
-  }, [location.pathname]);
-
-  return null;
-};
-
-const parseDisclosureInput = (input: string) => {
-  const trimmed = input.trim();
-  if (!trimmed) {
-    throw new Error('Disclosure JSON is required.');
-  }
-  const parseJson = (value: string) => JSON.parse(value);
-  try {
-    const parsed = parseJson(trimmed);
-    if (typeof parsed === 'string') {
-      return parseJson(parsed);
-    }
-    return parsed;
-  } catch (error) {
-    const firstBrace = trimmed.indexOf('{');
-    const lastBrace = trimmed.lastIndexOf('}');
-    if (firstBrace !== -1 && lastBrace > firstBrace) {
-      const candidate = trimmed.slice(firstBrace, lastBrace + 1);
-      try {
-        return parseJson(candidate);
-      } catch {
-      }
-    }
-    const message = error instanceof Error ? error.message : 'Invalid disclosure JSON.';
-    throw new Error(`Disclosure JSON is invalid. Paste the full package JSON. ${message}`);
-  }
-};
-
-const parseTimestamp = (value?: string | null) => {
-  if (!value) {
-    return 0;
-  }
-  const parsed = Date.parse(value);
-  return Number.isNaN(parsed) ? 0 : parsed;
-};
-
-const formatTimestamp = (value?: string | null) => {
-  if (!value) {
-    return 'N/A';
-  }
-  const parsed = Date.parse(value);
-  if (Number.isNaN(parsed)) {
-    return value;
-  }
-  return new Date(parsed).toLocaleString();
-};
-
-const activityStatusSymbols: Record<ActivityStatus, string> = {
-  processing: '...',
-  pending: '...',
-  confirmed: '✓',
-  failed: 'X'
-};
-
-const activityStatusLabels: Record<ActivityStatus, string> = {
-  processing: 'Processing',
-  pending: 'Pending',
-  confirmed: 'Confirmed',
-  failed: 'Failed'
-};
-
-const activityStatusClasses: Record<ActivityStatus, string> = {
-  processing: 'border-ionosphere/40 text-ionosphere bg-ionosphere/10',
-  pending: 'border-ionosphere/30 text-ionosphere/80 bg-ionosphere/5',
-  confirmed: 'border-ionosphere/40 text-ionosphere bg-ionosphere/15',
-  failed: 'border-guard/40 text-guard bg-guard/10'
-};
-
-const deriveHttpUrl = (wsUrl: string, httpUrl?: string) => {
-  const trimmedWsUrl = wsUrl.trim();
-  const wsAsHttp =
-    trimmedWsUrl.startsWith('ws://')
-      ? `http://${trimmedWsUrl.slice('ws://'.length)}`
-      : trimmedWsUrl.startsWith('wss://')
-        ? `https://${trimmedWsUrl.slice('wss://'.length)}`
-        : trimmedWsUrl;
-
-  const parseEndpoint = (value: string) => {
-    try {
-      return new URL(value);
-    } catch {
-      return null;
-    }
-  };
-  const isLoopbackHost = (host: string) => host === '127.0.0.1' || host === 'localhost' || host === '::1';
-  const wsEndpoint = parseEndpoint(trimmedWsUrl);
-
-  if (httpUrl && httpUrl.trim()) {
-    const trimmedHttpUrl = httpUrl.trim();
-    const httpEndpoint = parseEndpoint(trimmedHttpUrl);
-
-    // Local profiles frequently drift when only one endpoint is edited.
-    if (
-      wsEndpoint &&
-      httpEndpoint &&
-      isLoopbackHost(wsEndpoint.hostname) &&
-      isLoopbackHost(httpEndpoint.hostname) &&
-      wsEndpoint.port &&
-      httpEndpoint.port &&
-      wsEndpoint.port !== httpEndpoint.port
-    ) {
-      return wsAsHttp;
-    }
-
-    return trimmedHttpUrl;
-  }
-
-  return wsAsHttp;
-};
-
-const parsePortFromUrl = (value?: string | null): number | undefined => {
-  if (!value) {
-    return undefined;
-  }
-  const trimmed = value.trim();
-  if (!trimmed) {
-    return undefined;
-  }
-  try {
-    const parsed = new URL(trimmed);
-    if (!parsed.port) {
-      return undefined;
-    }
-    const port = Number.parseInt(parsed.port, 10);
-    if (!Number.isInteger(port) || port < 1 || port > 65535) {
-      return undefined;
-    }
-    return port;
-  } catch {
-    return undefined;
-  }
-};
-
-const normalizeRpcPort = (value?: number): number | undefined => {
-  if (typeof value !== 'number' || !Number.isInteger(value)) {
-    return undefined;
-  }
-  if (value < 1 || value > 65535) {
-    return undefined;
-  }
-  return value;
-};
-
-const parseListenAddrPort = (value?: string | null): number | undefined => {
-  if (!value) {
-    return undefined;
-  }
-  const parts = value.trim().split('/');
-  for (let index = 0; index < parts.length - 1; index += 1) {
-    if (parts[index] !== 'tcp') {
-      continue;
-    }
-    const port = Number.parseInt(parts[index + 1], 10);
-    if (Number.isInteger(port) && port >= 1 && port <= 65535) {
-      return port;
-    }
-  }
-  return undefined;
-};
-
-const rewriteListenAddrPort = (value: string, port: number): string => {
-  const parts = value.trim().split('/');
-  for (let index = 0; index < parts.length - 1; index += 1) {
-    if (parts[index] === 'tcp') {
-      parts[index + 1] = String(port);
-      return parts.join('/');
-    }
-  }
-  return value;
-};
-
-const isLoopbackWsEndpoint = (value: string): boolean =>
-  value.startsWith('ws://127.0.0.1:') || value.startsWith('ws://localhost:') || value.startsWith('ws://[::1]:');
-
-const isLoopbackHttpEndpoint = (value?: string): boolean =>
-  Boolean(
-    value &&
-      (value.startsWith('http://127.0.0.1:') ||
-        value.startsWith('http://localhost:') ||
-        value.startsWith('http://[::1]:') ||
-        value.startsWith('https://127.0.0.1:') ||
-        value.startsWith('https://localhost:') ||
-        value.startsWith('https://[::1]:'))
-  );
-
-const rewriteLoopbackWsEndpoint = (value: string, port: number): string => {
-  if (value.startsWith('ws://localhost:')) {
-    return `ws://localhost:${port}`;
-  }
-  if (value.startsWith('ws://[::1]:')) {
-    return `ws://[::1]:${port}`;
-  }
-  return `ws://127.0.0.1:${port}`;
-};
-
-const rewriteLoopbackHttpEndpoint = (value: string | undefined, port: number): string => {
-  if (!value) {
-    return `http://127.0.0.1:${port}`;
-  }
-  if (value.startsWith('http://localhost:')) {
-    return `http://localhost:${port}`;
-  }
-  if (value.startsWith('http://[::1]:')) {
-    return `http://[::1]:${port}`;
-  }
-  if (value.startsWith('https://localhost:')) {
-    return `https://localhost:${port}`;
-  }
-  if (value.startsWith('https://[::1]:')) {
-    return `https://[::1]:${port}`;
-  }
-  return `http://127.0.0.1:${port}`;
-};
-
-const inferRpcPort = (connection: NodeConnection): number =>
-  normalizeRpcPort(connection.rpcPort) ??
-  parsePortFromUrl(connection.wsUrl) ??
-  parsePortFromUrl(connection.httpUrl) ??
-  defaultRpcPort;
-
-const normalizeLocalConnectionEndpoints = (connection: NodeConnection): NodeConnection => {
-  if (connection.mode !== 'local') {
-    return connection;
-  }
-
-  const rpcPort = inferRpcPort(connection);
-  const updates: Partial<NodeConnection> = {};
-
-  if (normalizeRpcPort(connection.rpcPort) !== rpcPort) {
-    updates.rpcPort = rpcPort;
-  }
-
-  if (!connection.wsUrl?.trim()) {
-    updates.wsUrl = `ws://127.0.0.1:${rpcPort}`;
-  } else if (isLoopbackWsEndpoint(connection.wsUrl) && parsePortFromUrl(connection.wsUrl) !== rpcPort) {
-    updates.wsUrl = rewriteLoopbackWsEndpoint(connection.wsUrl, rpcPort);
-  }
-
-  const effectiveWsUrl = updates.wsUrl ?? connection.wsUrl;
-  const derivedHttpUrl = deriveHttpUrl(effectiveWsUrl, connection.httpUrl);
-  if (!connection.httpUrl?.trim()) {
-    updates.httpUrl = derivedHttpUrl;
-  } else if (isLoopbackHttpEndpoint(connection.httpUrl) && parsePortFromUrl(connection.httpUrl) !== rpcPort) {
-    updates.httpUrl = rewriteLoopbackHttpEndpoint(connection.httpUrl, rpcPort);
-  }
-
-  if (Object.keys(updates).length === 0) {
-    return connection;
-  }
-  return { ...connection, ...updates };
-};
-
-const isRoutineNetworkRetryLog = (line: string) =>
-  /failed to connect to peer|handshake failed|rate-limited peer address announcement/i.test(line);
-
-const classifyLogLevel = (line: string): LogLevel => {
-  if (/\bWARN\b|\bWarning\b/i.test(line)) {
-    return 'warn';
-  }
-  if (isRoutineNetworkRetryLog(line)) {
-    return 'warn';
-  }
-  if (/\bERROR\b|\bpanic\b/i.test(line)) {
-    return 'error';
-  }
-  if (/\bDEBUG\b/i.test(line)) {
-    return 'debug';
-  }
-  return 'info';
-};
-
-const classifyLogCategory = (line: string): LogCategory => {
-  if (/mining|nonce|hashrate|POW|coinbase/i.test(line)) {
-    return 'mining';
-  }
-  if (/imported block|block imported|sync|syncing/i.test(line)) {
-    return 'sync';
-  }
-  if (/peer|network|p2p|seed|broadcast/i.test(line)) {
-    return 'network';
-  }
-  if (/storage|state_root|nullifier_root|da_root/i.test(line)) {
-    return 'storage';
-  }
-  if (/rpc|jsonrpc|http/i.test(line)) {
-    return 'rpc';
-  }
-  if (/consensus|execute_extrinsics|block builder|block built/i.test(line)) {
-    return 'consensus';
-  }
-  return 'other';
-};
-
-const highlightLog = (line: string) => {
-  const routineNetworkRetry = isRoutineNetworkRetryLog(line);
-  if (/Block mined/i.test(line)) {
-    return 'Block mined';
-  }
-  if (/Block imported successfully|Block imported/i.test(line)) {
-    return 'Block imported';
-  }
-  if (/sync complete/i.test(line)) {
-    return 'Sync complete';
-  }
-  if ((/\bWARN\b|\bWarning\b/i.test(line)) && !routineNetworkRetry) {
-    return 'Warning';
-  }
-  if ((/\bERROR\b|\bpanic\b/i.test(line)) && !routineNetworkRetry) {
-    return 'Error';
-  }
-  return undefined;
-};
-
-const formatLogTimestamp = (value: string) => {
-  const parsed = new Date(value);
-  if (!Number.isNaN(parsed.getTime())) {
-    return parsed.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit', second: '2-digit', hour12: false });
-  }
-  return value;
-};
-
-const parseLogLine = (line: string, index: number): LogEntry => {
-  const timestampMatch =
-    line.match(/^(\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?Z)\s+(.*)$/) ??
-    line.match(/^(\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2})\s+(.*)$/);
-  const timestamp = timestampMatch ? formatLogTimestamp(timestampMatch[1]) : null;
-  const message = (timestampMatch ? timestampMatch[2] : line).replace(/^(TRACE|DEBUG|INFO|WARN|ERROR)\s+/, '');
-  return {
-    id: `${index}-${message.slice(0, 12)}`,
-    timestamp,
-    level: classifyLogLevel(line),
-    category: classifyLogCategory(line),
-    message,
-    raw: line,
-    highlight: highlightLog(line)
-  };
-};
-
-const buildDefaultConnection = (): NodeConnection => ({
-  id: makeId(),
-  label: defaultDevConnectionLabel,
-  mode: 'local',
-  participationRole: 'full_node',
-  wsUrl: `ws://127.0.0.1:${defaultRpcPort}`,
-  httpUrl: `http://127.0.0.1:${defaultRpcPort}`,
-  dev: true,
-  tmp: false,
-  basePath: defaultDevBasePath,
-  rpcPort: defaultRpcPort,
-  p2pPort: defaultP2pPort,
-  mineThreads: defaultMineThreads,
-  miningIntent: false,
-  ciphertextDaRetentionBlocks: 0,
-  proofDaRetentionBlocks: 0,
-  daStoreCapacity: 1024,
-  rpcMethods: 'unsafe',
-  rpcCorsAll: false,
-  seeds: approvedSeeds,
-  maxPeers: 50
-});
-
-const buildDefaultConnections = () => [buildDefaultConnection()];
-
-const normalizeRpcControlPlane = (connection: NodeConnection): NodeConnection => {
-  if (connection.mode !== 'local') {
-    return connection.rpcMethods === 'safe' ? connection : { ...connection, rpcMethods: 'safe' };
-  }
-  if (connection.rpcExternal) {
-    return connection.rpcMethods === 'safe' ? connection : { ...connection, rpcMethods: 'safe' };
-  }
-  return connection.rpcMethods === 'unsafe' ? connection : { ...connection, rpcMethods: 'unsafe' };
-};
-
-const normalizeConnection = (connection: NodeConnection): NodeConnection => {
-  const {
-    operatorEndpoint: _operatorEndpoint,
-    workerName: _workerName,
-    payoutAddress: _payoutAddress,
-    poolAuthToken: _poolAuthToken,
-    poolShareBits: _poolShareBits,
-    ...sanitizedConnection
-  } = connection as NodeConnection & {
-    operatorEndpoint?: string;
-    workerName?: string;
-    payoutAddress?: string;
-    poolAuthToken?: string;
-    poolShareBits?: number;
-  };
-  const hasDefaultLocalLabel = legacyDefaultConnectionLabels.has(sanitizedConnection.label);
-  const hasLegacyHegemonLabel = legacyHegemonConnectionLabels.has(sanitizedConnection.label?.trim());
-  const isDefaultLocal =
-    sanitizedConnection.mode === 'local' &&
-    hasDefaultLocalLabel &&
-    (!sanitizedConnection.basePath ||
-      sanitizedConnection.basePath === '~/.hegemon-node' ||
-      sanitizedConnection.basePath === defaultDevBasePath);
-  const isDefaultTestnet =
-    sanitizedConnection.mode === 'local' &&
-    sanitizedConnection.label === 'Testnet node' &&
-    (!sanitizedConnection.basePath || sanitizedConnection.basePath === '~/.hegemon-node-testnet');
-
-  let next = normalizeLocalConnectionEndpoints(sanitizedConnection);
-  const isLegacyHegemonLocalProfile =
-    sanitizedConnection.mode === 'local' &&
-    hasLegacyHegemonLabel &&
-    Boolean(sanitizedConnection.dev) &&
-    inferRpcPort(next) === defaultRpcPort &&
-    isLoopbackWsEndpoint(next.wsUrl) &&
-    (!next.httpUrl || isLoopbackHttpEndpoint(next.httpUrl));
-  if ((isDefaultLocal || isLegacyHegemonLocalProfile) && next.label !== defaultDevConnectionLabel) {
-    next = { ...next, label: defaultDevConnectionLabel };
-  }
-  if (isDefaultLocal && (!next.basePath || next.basePath === '~/.hegemon-node')) {
-    next = { ...next, basePath: defaultDevBasePath };
-  }
-  if ((isDefaultLocal || isDefaultTestnet) && (next.mineThreads ?? defaultMineThreads) > maxDesktopMineThreads) {
-    next = { ...next, mineThreads: maxDesktopMineThreads };
-  }
-  if (
-    isDefaultLocal &&
-    inferRpcPort(next) === legacyDesktopRpcPort &&
-    isLoopbackWsEndpoint(next.wsUrl) &&
-    (!next.httpUrl || isLoopbackHttpEndpoint(next.httpUrl))
-  ) {
-    next = {
-      ...next,
-      rpcPort: defaultRpcPort,
-      wsUrl: rewriteLoopbackWsEndpoint(next.wsUrl, defaultRpcPort),
-      httpUrl: rewriteLoopbackHttpEndpoint(next.httpUrl, defaultRpcPort)
-    };
-  }
-  const inferredParticipationRole = inferParticipationRole(next);
-  if (next.participationRole !== inferredParticipationRole) {
-    next = { ...next, participationRole: inferredParticipationRole };
-  }
-  if (next.mode === 'remote' && !next.httpUrl?.trim()) {
-    next = { ...next, httpUrl: deriveHttpUrl(next.wsUrl) };
-  }
-  next = normalizeRpcControlPlane(next);
-
-  const currentSeeds = (next.seeds ?? '').trim();
-  const normalizedSeeds = normalizeSeedsValue(next.seeds);
-  if (currentSeeds && normalizedSeeds !== currentSeeds.toLowerCase()) {
-    next = { ...next, seeds: normalizedSeeds };
-  }
-  if (isDefaultLocal || isDefaultTestnet) {
-    if (normalizedSeeds !== approvedSeeds) {
-      next = { ...next, seeds: approvedSeeds };
-    }
-  }
-
-  if (next.mode === 'local') {
-    if (next.p2pPort === 31333) {
-      next = { ...next, p2pPort: canonicalTestnetP2pPort };
-    }
-    const listenAddrPort = parseListenAddrPort(next.listenAddr);
-    if (listenAddrPort === 31333 && next.listenAddr) {
-      next = {
-        ...next,
-        listenAddr: rewriteListenAddrPort(next.listenAddr, canonicalTestnetP2pPort)
-      };
-    }
-  }
-
-  if ((isDefaultLocal || isDefaultTestnet) && (!next.mineThreads || next.mineThreads === 1)) {
-    next = { ...next, mineThreads: defaultMineThreads };
-  }
-
-  if (inferredParticipationRole !== 'authoring_pool' && next.miningIntent) {
-    next = { ...next, miningIntent: false };
-  }
-
-  return next;
-};
-
-const shouldAutoStartDefaultProfile = (connection: NodeConnection): boolean => {
-  const normalized = normalizeLocalConnectionEndpoints(connection);
-  const role = inferParticipationRole(normalized);
-  const miningProfileAllowed =
-    role === 'authoring_pool' &&
-    Boolean(normalized.miningIntent) &&
-    !validateShieldedAddressInput(normalized.minerAddress ?? '', 'Miner address');
-  const relayProfileAllowed = role === 'full_node' && !normalized.miningIntent;
-  return (
-    normalized.mode === 'local' &&
-    normalized.label === defaultDevConnectionLabel &&
-    Boolean(normalized.dev) &&
-    !normalized.tmp &&
-    !normalized.rpcExternal &&
-    (relayProfileAllowed || miningProfileAllowed) &&
-    inferRpcPort(normalized) === defaultRpcPort &&
-    normalized.basePath === defaultDevBasePath &&
-    isLoopbackWsEndpoint(normalized.wsUrl) &&
-    (!normalized.httpUrl || isLoopbackHttpEndpoint(normalized.httpUrl)) &&
-    normalizeSeedsValue(normalized.seeds) === approvedSeeds
-  );
-};
-
-const findDefaultManagedConnection = (connections: NodeConnection[]) =>
-  connections.find((connection) => shouldAutoStartDefaultProfile(connection)) ?? null;
+import {
+  activeConnectionKey,
+  approvedSeedEntries,
+  approvedSeeds,
+  blockAlertEnabledKey,
+  canonicalTestnetP2pPort,
+  connectionModeLabels,
+  connectionsKey,
+  defaultDevBasePath,
+  defaultDevConnectionLabel,
+  defaultMineThreads,
+  defaultP2pPort,
+  defaultRpcPort,
+  defaultStorePath,
+  hegemonNetworkName,
+  hegemonNetworkVersionLabel,
+  inferParticipationRole,
+  legacyDefaultConnectionLabels,
+  legacyDesktopRpcPort,
+  legacyHegemonConnectionLabels,
+  legacySeedAliases,
+  maxDesktopMineThreads,
+  minWalletPassphraseLength,
+  participationRoleLabels,
+  participationRoleMeta,
+  shieldedAddressDataCharset,
+  shieldedAddressLength,
+  shieldedAddressPrefix,
+  shieldedAddressSeparatorPattern,
+  walletAutoLockEnabledKey,
+  walletAutoLockMinutesKey,
+  walletConnectionKey,
+  walletHistoryPageSize,
+  walletStorePathKey
+} from './lib/config';
+import {
+  activityStatusClasses,
+  activityStatusLabels,
+  activityStatusSymbols,
+  buildBlockAlertPattern,
+  clampAutoLockMinutes,
+  formatAddress,
+  formatBlockCount,
+  formatBytes,
+  formatCompactPath,
+  formatDuration,
+  formatEndpoint,
+  formatHash,
+  formatHashRate,
+  formatHgm,
+  formatNumber,
+  formatSeedList,
+  formatTimestamp,
+  humanizeWalletAddressError,
+  isWalletSessionClosedError,
+  makeId,
+  normalizeShieldedAddressInput,
+  normalizeTxId,
+  parseDisclosureInput,
+  parseTimestamp,
+  toBaseUnits,
+  validateShieldedAddressInput
+} from './lib/format';
+import {
+  buildDefaultConnection,
+  buildDefaultConnections,
+  canonicalizeSeedEntry,
+  deriveHttpUrl,
+  findDefaultManagedConnection,
+  inferRpcPort,
+  isLoopbackHttpEndpoint,
+  isLoopbackWsEndpoint,
+  normalizeConnection,
+  normalizeLocalConnectionEndpoints,
+  normalizeNetworkDisplayName,
+  normalizeRpcControlPlane,
+  normalizeRpcPort,
+  normalizeSeedsValue,
+  parseListenAddrPort,
+  parsePortFromUrl,
+  rewriteListenAddrPort,
+  rewriteLoopbackHttpEndpoint,
+  rewriteLoopbackWsEndpoint,
+  shouldAutoStartDefaultProfile
+} from './lib/connections';
+import {
+  classifyLogCategory,
+  classifyLogLevel,
+  formatLogTimestamp,
+  highlightLog,
+  isRoutineNetworkRetryLog,
+  logCategoryLabels,
+  logCategoryOrder,
+  parseLogLine
+} from './lib/logs';
+import type {
+  ActivityEntry,
+  ActivityStatus,
+  ActivityStep,
+  AppIconName,
+  BlockAlertStep,
+  BlockAlertTone,
+  DisclosureGroup,
+  EmptyStateIconName,
+  LogCategory,
+  LogEntry,
+  LogLevel,
+  NodeTransition,
+  NodeTransitionAction,
+  SendAttempt,
+  UiTone
+} from './lib/appTypes';
+import { AppIcon } from './components/AppIcon';
+import { EmptyStateIcon } from './components/EmptyStateIcon';
+import { ScrollToTop } from './components/ScrollToTop';
 
 export default function App() {
   const [connections, setConnections] = useState<NodeConnection[]>([]);
